+++ source/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.9.4
+
+Unreleased
+
 # 2.9.3
 
 Andreas Abel <andreas.abel@gu.se>  September 2021